api_clip_embeddings/requirements.txt
flask
torch
torchvision
transformers
requests
pillow
scikit-learn
matplotlib
pandas